requirements.txt
openai>=1.0.0
anthropic>=0.8.0
google-generativeai>=0.3.0
mistralai>=0.0.7
tiktoken>=0.5.0
numpy>=1.24.0
aiohttp>=3.8.0
pydantic>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
plotly>=5.17.0
pandas>=2.0.0
redis>=5.0.0
prometheus-client>=0.19.0
click>=8.1.0
rich>=13.0.0
pyyaml>=6.0
streamlit>=1.28.0
matplotlib>=3.7.0
seaborn>=0.12.0
